Refactor OauthGoogle.js
Clean up responseGoogle to only use addUser() rather than test() as well
Remove test() and put .then statement on end of addUser() for easier
readability and more concise code.

Use the useSelector hook to return login status...If not logged in then
the login or signup with Google button shows on page.

If user is logged in then the logout button renders.

Add logout component to OauthGoogle component return page rather than on App.js

Remove logout component from App.js

Pass entire user to loginReducer rather than just the id for more concise coding
when using that state in other areas of the application.

Add .data to reducers to adapt to the way the info is being passed to the
reducers compared to before